+++ client/src/app/(withauth)/dashboard/page.tsx
@@ -11,7 +11,6 @@ import ProductCard from "@/components/productcard-component"
 import { GlobalContext } from "@/context/globalstate.provider"
 import Error from "@/components/error-component"
 import { uiConstants } from "@/constants/global-constants"
-import SubHeader from "@/components/sub-header-component"
 import { BellIcon, CubeIcon } from "@radix-ui/react-icons"
 import { useRouter } from "next/navigation"
 import Hero from "@/components/hero-component"
@@ -68,14 +67,6 @@ export default function Page() {
   return (
     <Fragment>
       <Show when={!products.isLoading && !activities.isLoading}>
-        <SubHeader>
-          <div className="d-flex justify-content-between align-items-center">
-            <h4 className="text-white">Dashboard</h4>
-            <div className="ml-auto">
-              <CubeIcon className="icon-subheader" onClick={() => router.push("/workspace")} />
-            </div>
-          </div>
-        </SubHeader>
         <Container>
           <Row>
             <Col xl={9} lg={12} md={12} sm={12} xs={12}>